Added Id field to registry Entry
