+++ Registry.Web.Data/Models/Batch.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Data;
 using System.Text;
-using Registry.Common;
 
 namespace Registry.Web.Data.Models
 {
@@ -32,24 +31,4 @@ namespace Registry.Web.Data.Models
     {
         Running, Committed, Rolledback
     }
-
-
-    public class Entry
-    {
-        [Key]
-        public string Path { get; set; }
-        
-        [Required]
-        public string Hash { get; set; }
-        [Required]
-        public EntryType Type { get; set; }
-        [Required]
-        public int Size { get; set; }
-        [Required]
-        public DateTime AddedOn { get; set; }
-
-        [Required]
-        public Batch Batch { get; set; }
-    }
-
 }
